add update data function inside ViewModel

+++ ZalZaZa/ZalZaZa/Models/Contestant.swift
@@ -13,8 +13,8 @@ struct Contestant: Identifiable, Decodable {
     var id:String  // Assume id is given when data get fetched
     var pw:String?
     var name:String
-    var height:Float?
-    var weight:Float?
+    var height:Double?
+    var weight:Double?
     var age:Int?
     var image:String?
     var birthdate:Date?
